--- a/src/Hashids.net/Hashids.net.csproj
+++ b/src/Hashids.net/Hashids.net.csproj
@@ -18,8 +18,14 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.ObjectPool">
+          <Version>5.0.5</Version>
+        </PackageReference>
         <PackageReference Include="System.Buffers">
             <Version>4.5.1</Version>
+        </PackageReference>
+        <PackageReference Include="System.Memory">
+          <Version>4.5.4</Version>
         </PackageReference>
     </ItemGroup>