--- a/src/Hashids.net/IHashids.cs
+++ b/src/Hashids.net/IHashids.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace HashidsNet
 {