latest benchmarks, reduced some allocations and improved runtime

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET Core SDK=5.0.202
  [Host]     : .NET Core 5.0.5 (CoreCLR 5.0.521.16609, CoreFX 5.0.521.16609), X64 RyuJIT
  DefaultJob : .NET Core 5.0.5 (CoreCLR 5.0.521.16609, CoreFX 5.0.521.16609), X64 RyuJIT


```
|         Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|  RoundtripInts | 4.250 μs | 0.0814 μs | 0.1000 μs | 0.2213 |     - |     - |   1.85 KB |
| RoundtripLongs | 4.758 μs | 0.0574 μs | 0.0537 μs | 0.2518 |     - |     - |    2.1 KB |
|   RoundtripHex | 4.388 μs | 0.0196 μs | 0.0184 μs | 0.3815 |     - |     - |   3.16 KB |


|         Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|  RoundtripInts | 4.301 μs | 0.0545 μs | 0.0510 μs | 0.1755 |     - |     - |   1.45 KB |
| RoundtripLongs | 4.883 μs | 0.0391 μs | 0.0366 μs | 0.2060 |     - |     - |    1.7 KB |
|   RoundtripHex | 4.443 μs | 0.0173 μs | 0.0153 μs | 0.3052 |     - |     - |   2.55 KB |


|         Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|  RoundtripInts | 4.122 us | 0.0274 us | 0.0257 us | 0.1755 |     - |     - |   1.45 KB |
| RoundtripLongs | 4.699 us | 0.0880 us | 0.0780 us | 0.2060 |     - |     - |    1.7 KB |
|   RoundtripHex | 4.300 us | 0.0607 us | 0.0538 us | 0.2899 |     - |     - |   2.41 KB |


|         Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|  RoundtripInts | 4.119 us | 0.0102 us | 0.0090 us | 0.1755 |     - |     - |   1.45 KB |
| RoundtripLongs | 4.608 us | 0.0888 us | 0.0912 us | 0.2060 |     - |     - |    1.7 KB |
|   RoundtripHex | 4.314 us | 0.0518 us | 0.0485 us | 0.2899 |     - |     - |   2.41 KB |